rename class and variables
