+++ domain/src/main/kotlin/com/minkyu/order/domain/Order.kt
@@ -1,6 +1,13 @@
 package com.minkyu.order.domain
 
+import java.math.BigDecimal
+
 data class Order(
     val orderId: OrderId? = null,
-    val orderLine: OrderLine
-)
+    val orderLine: OrderLine,
+    var totalPrice: BigDecimal? = null
+) {
+    init {
+        this.totalPrice = orderLine.getTotalPrice()
+    }
+}
