+++ domain/src/main/kotlin/com/minkyu/order/domain/OrderItem.kt
@@ -3,6 +3,6 @@ package com.minkyu.order.domain
 data class OrderItem(
     val orderItemId: Long? = null,
     val orderItemName: String,
-    val amount: Int,
+    val quantity: Int,
     val price: Price
 )
